Restructure app with routes and modular predict endpoint
